// jovo-platforms/jovo-platform-googleassistant/src/conversational-actions/core/Interfaces.ts
export type SlotFillingStatus =
  | 'UNSPECIFIED'
  | 'INITIALIZED'
  | 'COLLECTING'
  | 'FINALIZING'
  | 'FINAL';
export type SlotMode = 'MODE_UNSPECIFIED' | 'OPTIONAL' | 'REQUIRED';
export type SlotStatus = 'SLOT_UNSPECIFIED' | 'EMPTY' | 'INVALID' | 'FILLED';
// ...
export type TypeOverrideMode = 'TYPE_UNSPECIFIED' | 'TYPE_REPLACE' | 'TYPE_MERGE';
export type Capability =
  | 'UNSPECIFIED'
  | 'SPEECH'
  | 'RICH_RESPONSE'
  | 'LONG_FORM_AUDIO'
  | 'INTERACTIVE_CANVAS'
  | 'WEB_LINK'
  | 'HOME_STORAGE';
export type AccountLinkingStatus = 'ACCOUNT_LINKING_STATUS_UNSPECIFIED' | 'NOT_LINKED' | 'LINKED';
export type UserVerificationStatus = 'USER_VERIFICATION_STATUS_UNSPECIFIED' | 'GUEST' | 'VERIFIED';
// ...
export interface Params {
  [key: string]: any;
}

export interface Handler {
  name?: string;
}
export interface Intent {
  name: string;
  params: Record<string, IntentParameterValue>;
  query?: string;
}

export interface IntentParameterValue {
  original: string;
  resolved: string;
}

export interface Simple {
  speech?: string;
  text?: string;
}

export interface Image {
  url: string;
  alt: string;
  height?: number;
  width?: number;
}
export interface OpenUrl {
  url: string;
  hint: UrlHint;
}

export interface Link {
  name: string;
}
export interface Card {
  title: string;
  subtitle?: string;
  text: string; // required, unless image is present
  image?: Image;
  imageFill?: ImageFill;
  button?: Link;
}

export interface TableColumn {
  header: string;
  align: HorizontalAlignment;
}

export interface TableCell {
  text: string;
}

export interface TableRow {
  cells: TableCell[];
  divider: boolean;
}

export interface Table {
  title: string;
  subtitle?: string;
  image?: Image;
  columns: TableColumn[];
  rows: TableRow[];
  button?: Link;
}

export interface MediaImage {
  large: Image;
  icon: Image;
}
export interface MediaObject {
  name: string;
  description: string;
  url: string;
  image: MediaImage;
  startOffset: string; // Example: "3.5s".
  optionalMediaControls?: OptionalMediaControls;
}

export interface Media {
  mediaType: MediaType;
  mediaObjects: MediaObject[];
}

export interface ExpectedPhrase {
  phrase: string;
  alternativePhrases: string[];
}

export interface ContinuousMatchConfig {
  expectedPhrases: ExpectedPhrase[];
  durationSeconds: number;
}

export interface Canvas {
  url: string;
  data: any;
  suppressMic?: boolean;
  continuousMatchConfig?: ContinuousMatchConfig;
  sendStateDataToCanvasApp: boolean;
}

export interface CollectionItem {
  key: string;
}
export interface Collection {
  title?: string;
  subtitle?: string;
  items?: CollectionItem[]; // min 2, max 10
  imageFill?: ImageFill;
}

export interface Content {
  // can only be one of the following
  card?: Card;
  image?: Image;
  table?: Table;
  media?: Media;
  canvas?: Canvas;
  collection?: Collection;
  list?: List;
}
export interface ListItem {
  key: string;
}
export interface List {
  title?: string;
  subtitle?: string;
  items?: ListItem[]; // min 2, max 30
}
export interface Suggestion {
  title: string;
}

export interface Prompt {
  append?: boolean; // deprecated
  override?: boolean;
  firstSimple?: Simple;
  content?: Content;
  lastSimple?: Simple;
  suggestions?: Suggestion[];
  link?: Link;
  canvas?: Canvas;
  orderUpdate?: OrderUpdate;
}

export interface Slot {
  mode: SlotMode;
  status: SlotStatus;
  value: any;
  updated: boolean;
  prompt?: Prompt;
}

export interface NextScene {
  name: string;
}

export interface Scene {
  name: string;
  slotFillingStatus: SlotFillingStatus;
  slots: Record<string, Slot>;
  next: NextScene;
}

export interface EntryDisplay {
  title: string;
  description?: string;
  image?: Image;
  footer?: string;
  openUrl?: OpenUrl;
}

export interface Entry {
  name: string;
  synonyms: string[];
  display?: EntryDisplay;
}

export interface SynonymType {
  entries: Entry[];
}

export interface TypeOverride {
  name: string;
  mode: TypeOverrideMode;
  synonym: SynonymType;
}

export interface Session {
  id: string;
  params: Params;
  typeOverrides?: TypeOverride[];
  languageCode?: string;
}
export interface Home {
  params: Params;
}

export interface Device {
  capabilities: Capability[];
}
export interface MediaContext {
  progress: string; // Example: "3.5s
}
export interface User {
  locale: string;
  params: Params;
  accountLinkingStatus: AccountLinkingStatus;
  verificationStatus: UserVerificationStatus;
  lastSeenTime: string;
  engagement?: Engagement;
  packageEntitlements?: PackageEntitlement[];
}

export interface Engagement {
  pushNotificationIntents?: IntentSubscription[];
  dailyUpdateIntents?: IntentSubscription[];
}
export interface IntentSubscription {
  intent: string;
  contentTitle: string;
}

export interface PackageEntitlement {
  packageName: string;
  entitlements: Entitlement[];
}

export interface Entitlement {
  sku: string;
  skuType: SkuType;
  inAppDetails: SignedData;
}

export interface SignedData {
  // tslint:disable-next-line:no-any
  inAppPurchaseData: any;
  inAppDataSignature: string;
}
export interface Context {
  media: MediaContext;
}

export interface OrderUpdate {
  type?: string;
  reason?: string;
  order: Order;
  updateMask: string;
  userNotification?: UserNotification;
}
export interface UserNotification {
  title: string;
  text: string;
}
export interface Order {
  googleOrderId?: string;
  merchantOrderId?: string;
  userVisibleOrderId?: string;
  userVisibleStateLabel?: string;

  buyerInfo?: UserInfo;
  image?: Image;
  createTime?: string;
  lastUpdateTime?: string;
  transactionMerchant?: Merchant;
  contents?: Contents;
  priceAttributes?: PriceAttribute[];
  followUpActions?: Action[];
  paymentData?: PaymentData;
  termsOfServiceUrl?: string;
  note?: string;
  promotions?: Promotion[];
  disclosures?: Disclosure[];

  // purchase OR ticket
  purchase?: PurchaseOrderExtension;
  ticket?: TicketOrderExtension;
}